+++ fresh.gen.ts
@@ -10,8 +10,8 @@ import * as $3 from "./routes/docs.tsx";
 import * as $4 from "./routes/index.tsx";
 import * as $$0 from "./islands/CopyToClip.tsx";
 import * as $$1 from "./islands/CsvChart.tsx";
-import * as $$2 from "./islands/Header.tsx";
-import * as $$3 from "./islands/MemoryChart.tsx";
+import * as $$2 from "./islands/MemoryChart.tsx";
+import * as $$3 from "./islands/Nav.tsx";
 import * as $$4 from "./islands/csv.tsx";
 
 const manifest = {
@@ -25,8 +25,8 @@ const manifest = {
   islands: {
     "./islands/CopyToClip.tsx": $$0,
     "./islands/CsvChart.tsx": $$1,
-    "./islands/Header.tsx": $$2,
-    "./islands/MemoryChart.tsx": $$3,
+    "./islands/MemoryChart.tsx": $$2,
+    "./islands/Nav.tsx": $$3,
     "./islands/csv.tsx": $$4,
   },
   baseUrl: import.meta.url,
