+++ islands/CsvChart.tsx
@@ -1,118 +1,23 @@
 import * as chartjs from "https://cdn.jsdelivr.net/npm/chart.js@3.9.1/dist/chart.min.js";
-import { useEffect } from "preact/hooks";
-
+import { useEffect, useState } from "preact/hooks";
+import styles  from '../utils/styles.ts';
 
  export default function CsvChart(props: { data: [] | [string, number[]][] }) {
+  let x, rss, committed, heapTotal, heapUsed, external: number[] = [];
+  props.data.forEach(e => {
+    switch(e[0]){
+      case 'x': x = e[1]; break;
+      case 'rss': rss = e[1]; break;
+      case 'committed': committed = e[1]; break;
+      case 'heapTotal': heapTotal = e[1]; break;
+      case 'heapUsed': heapUsed = e[1]; break;
+      case 'external': external = e[1]; break;
+      default: throw new Error('Please check your csv columns. Your columns must be named x, rss, commmitted, heapTotal, heapUsed, and external.')
+    }
+  })
   if(props.data.length !== 0){
-    const chartStyle = {
-      labels: [...props.data[0][1]],
-      datasets: [
-        {
-          label: "RSS",
-          data: [...props.data[5][1]],
-          backgroundColor: [
-            "rgba(105, 0, 132, .2)",
-          ],
-          borderColor: [
-            "rgba(200, 99, 132, .7)",
-          ],
-          fill: true,
-          borderWidth: 1,
-          tension: 0.5,
-        },
-        {
-          label: "Committed Heap",
-          data: [...props.data[1][1]],
-          backgroundColor: [
-            "rgba(0, 20, 20, .2)",
-          ],
-          borderColor: [
-            "rgba(0, 30, 20, .7)",
-          ],
-          fill: true,
-          borderWidth: 1,
-        },
-        {
-          label: "Heap Total",
-          data: [...props.data[2][1]],
-          backgroundColor: [
-            "rgba(0, 137, 132, .2)",
-          ],
-          borderColor: [
-            "rgba(0, 10, 130, .7)",
-          ],
-          fill: true,
-          borderWidth: 1,
-        },
-        {
-          label: "Heap Used",
-          data: props.data[3][1],
-          backgroundColor: [
-            "rgba(0, 255, 0, .2)",
-          ],
-          borderColor: [
-            "rgba(0, 153, 0, .7)",
-          ],
-          fill: true,
-          borderWidth: 1,
-          tension: 0.5,
-        },
-        {
-          label: "External",
-          data: [...props.data[4][1]],
-          backgroundColor: [
-            "rgba(255, 102, 78, .2)",
-          ],
-          borderColor: [
-            "rgba(255, 0, 127, .7)",
-          ],
-          fill: true,
-          borderWidth: 1,
-          tension: 0.5,
-        },
-      ],
-    };
-  
-    const chartOptions = {
-      scales: {
-        yAxes: {
-          suggestedmax: 1000000,
-          suggestedmin: 0,
-          ticks: {
-            stepSize: 10000,
-          },
-          title: {
-            display: true,
-            text: 'Memory Usage (Kb)',
-            align: 'center',
-            padding: 16,
-            font: {
-              size: 24
-            }
-          }
-        },
-        xAxes: {
-          title: {
-            display: true,
-            text: "Data Points",
-            align: 'center',
-            padding: 12,
-            font: {
-              size: 24
-            }
-          }
-        }
-      },
-      animation: true,
-      elements: {
-        line: {
-        },
-        point: {
-          radius: 1
-        }
-      }
-    };
-  
+    const { chartStyle, chartOptions } = styles(x, rss, committed, heapTotal, heapUsed, external);
+
     useEffect(() => {
       const ctx1 = document.getElementById("myCsvChart");
       const lineChart = new chartjs.Chart(ctx1, {
@@ -120,7 +25,7 @@ import { useEffect } from "preact/hooks";
         data: chartStyle,
         options: chartOptions,
       });
-    }, [props.data]);
+    }, []);
 
     return(
