+++ islands/Nav.tsx
@@ -1,0 +1,47 @@
+
+import { useState, useEffect, useRef } from "preact/hooks";
+import MenuButton from "../components/MenuButton.tsx";
+import Links from "../components/Links.tsx";
+
+/**
+ * Site navigation links.
+ */
+export default function Nav(){
+  const ref = useRef(window);
+  const [navOpen, setNavOpen] = useState(false);
+  useEffect(() => {
+    let lastKnownWidth = 0;
+    let ticking = false;
+    const collapse = (width: number) => {
+      // console.log(width);
+      if (width > 768) {
+        setNavOpen(true);
+      } else {
+        setNavOpen(false);
+      }
+    };
+    const onResize = (e: Event) => {
+      lastKnownWidth = ref.current.innerWidth;
+      if (!ticking) {
+        ref.current.requestAnimationFrame(() => {
+          collapse(lastKnownWidth);
+          ticking = false;
+        });
+        ticking = true;
+      }
+    };
+    collapse(ref.current.innerWidth);
+    ref.current.addEventListener("resize", onResize);
+    return () => {
+      ref.current.removeEventListener("resize", onResize);
+    };
+  }, []);
+
+  
+    return (
+      <div class="flex flex-row justify-between items-center">
+        {navOpen ? <Links/> : <></>}
+        <MenuButton onClick = {() => setNavOpen(!navOpen)}/>
+      </div>
+  )
+}
