+++ routes/csv.tsx
@@ -1,5 +1,5 @@
 import UploadChart from "../islands/csv.tsx";
-import Header from "../islands/Header.tsx";
+import Header from "../components/Header.tsx";
 
 export default function Csv () {
