+++ routes/docs.tsx
@@ -1,4 +1,4 @@
-import Header from "../islands/Header.tsx";
+import Header from "../components/Header.tsx";
 import CopyToClip from "../islands/CopyToClip.tsx";
 
 export default function Docs() {
