+++ routes/index.tsx
@@ -1,7 +1,7 @@
 /** @jsx */
 
 import MemoryChart from "../islands/MemoryChart.tsx";
-import Header from "../islands/Header.tsx";
+import Header from "../components/Header.tsx";
 
 export default function Home() {
   console.log("hello from Home Export index.tsx");
